requirements.txt
requests
beautifulsoup4
gql[requests]
python-dotenv
schedule
psycopg2-binary
yfinance 